can delete address fields & addressbook entries

+++ Sql/dbo/Stored Procedures/AddressFields/AddressField_Delete.sql
@@ -1,8 +1,11 @@
 ﻿CREATE PROCEDURE [dbo].[AddressField_Delete]
+	@teamId int,
 	@fieldId int
 AS
-	DELETE FROM AddressFields_Bits WHERE fieldId=@fieldId
-	DELETE FROM AddressFields_DateTimes WHERE fieldId=@fieldId
-	DELETE FROM AddressFields_Numbers WHERE fieldId=@fieldId
-	DELETE FROM AddressFields_Text WHERE fieldId=@fieldId
-	DELETE FROM AddressFields WHERE fieldId=@fieldId
+	IF(SELECT COUNT(*) FROM AddressFields WHERE teamId=@teamId AND fieldId=@fieldId) > 0 BEGIN
+		DELETE FROM AddressFields_Bits WHERE fieldId=@fieldId
+		DELETE FROM AddressFields_DateTimes WHERE fieldId=@fieldId
+		DELETE FROM AddressFields_Numbers WHERE fieldId=@fieldId
+		DELETE FROM AddressFields_Text WHERE fieldId=@fieldId
+		DELETE FROM AddressFields WHERE fieldId=@fieldId
+	END
